+++ tcs-service/src/main/java/com/transformuk/hee/tis/tcs/service/dto/placementmanager/PlacementDTO.java
@@ -3,7 +3,6 @@ package com.transformuk.hee.tis.tcs.service.dto.placementmanager;
 import com.transformuk.hee.tis.tcs.api.enumeration.LifecycleState;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 
@@ -12,7 +11,7 @@ public class PlacementDTO implements Serializable {
   private Long id;
   private LocalDate dateFrom;
   private LocalDate dateTo;
-  private BigDecimal wte;
+  private Float wte;
   private String type;
   private PersonDTO trainee;
   private LifecycleState lifecycleState;
@@ -41,11 +40,11 @@ public class PlacementDTO implements Serializable {
     this.dateTo = dateTo;
   }
 
-  public BigDecimal getWte() {
+  public Float getWte() {
     return wte;
   }
 
-  public void setWte(BigDecimal wte) {
+  public void setWte(Float wte) {
     this.wte = wte;
   }
